requirements.txt
# Core FastAPI dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# PDF Processing
camelot-py[cv]==0.11.0
tabula-py==2.9.3
pdfplumber==0.11.4
PyPDF2==3.0.1

# OCR (Optional - install separately if needed)
# pytesseract==0.3.13
# Requires Tesseract-OCR binary: https://github.com/tesseract-ocr/tesseract

# Image Processing
opencv-python==4.10.0.84
Pillow==11.0.0

# Data Science
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1
matplotlib==3.10.0
openpyxl==3.1.5

# Ternary Plotting
python-ternary==1.0.8

# Additional Dependencies (auto-installed with above)
# These are listed for reference but don't need to be explicitly installed
# pydantic==2.10.6
# starlette==0.41.3
# click==8.1.8
# h11==0.14.0
